+++ cricket/cricket_enums.py
