+++ cricket/cricket_team.py
