+++ cricket/printer/fixture_list_type.py
