+++ cricket/reader/googlecalendar_utils.py
@@ -1,4 +1,4 @@
-from src.cricket_enums import FixtureType
+from cricket_enuns import FixtureType
 from datetime import datetime, timezone, date
 
 def clean_summary(summary):
