+++ tests/reader/test_read_googlecalendar_utils.py
@@ -1,4 +1,4 @@
-from src.cricket_enums import FixtureType
+from cricket_enums import FixtureType
 from datetime import datetime, timezone, date
 
 import pytest
